pyproject.toml
[project]
name = "shion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "google-genai>=1.12.1",
    "httpx[socks]>=0.28.1",
    "pydantic>=2.11.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "python-lsp-server[all]>=1.12.2",
    "ruff>=0.11.7",
]

[tool.black]
line-length=79
